+++ kukumo/kukumo-core/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog][1],
 and this project adheres to [Semantic Versioning][2].
 
+## [1.3.0] 2022-01-26
+### Added
+- Steps can use property substitution using the syntax`${property.name}`
+
 
 ## [1.2.1] 2021-10-22
 ### Fixed